feat(collect): split professional and daily recommendation (#396)
